+++ src/com/maddyhome/idea/vim/action/motion/leftright/MotionShiftRightAction.kt
@@ -45,7 +45,7 @@ class MotionShiftRightAction : VimCommandAction() {
 
     override fun motionWithoutKeyModel(editor: Editor, context: DataContext, cmd: Command) {
       editor.vimForEachCaret { caret ->
-        val newOffset = VimPlugin.getMotion().moveCaretToNextWord(editor, caret, cmd.count, false)
+        val newOffset = VimPlugin.getMotion().findOffsetOfNextWord(editor, caret.offset, cmd.count, false)
         MotionGroup.moveCaret(editor, caret, newOffset)
       }
     }