+++ src/com/maddyhome/idea/vim/action/motion/text/MotionBigWordRightAction.java
@@ -62,7 +62,7 @@ public class MotionBigWordRightAction extends MotionEditorAction {
       @Override
       public int getOffset(@NotNull Editor editor, @NotNull Caret caret, @NotNull DataContext context, int count,
       int rawCount, @Nullable Argument argument) {
-        return VimPlugin.getMotion().moveCaretToNextWord(editor, caret, count, true);
+        return VimPlugin.getMotion().findOffsetOfNextWord(editor, caret.getOffset(), count, true);
       }
     };
   }