+++ src/com/maddyhome/idea/vim/action/motion/text/MotionWordRightAction.kt
@@ -63,6 +63,6 @@ private object MotionWordRightActionHandler : MotionActionHandler.ForEachCaret()
                          count: Int,
                          rawCount: Int,
                          argument: Argument?): Int {
-    return VimPlugin.getMotion().moveCaretToNextWord(editor, caret, count, false)
+    return VimPlugin.getMotion().findOffsetOfNextWord(editor, caret.offset, count, false)
   }
 }